+++ take-screenshot/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
 
     defaultConfig {
-        applicationId "com.example.capture_screenshot_map"
+        applicationId "com.esri.arcgisruntime.sample.takescreenshot"
         minSdkVersion 16
         targetSdkVersion 23
         versionCode 1
