+++ take-screenshot/proguard-rules.pro
