+++ take-screenshot/src/main/res/values/colors.xml
