+++ take-screenshot/src/main/res/values/dimens.xml
