+++ take-screenshot/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">Capture Map Screenshot</string>
+    <string name="app_name">Take screenshot</string>
     <string name="menu_capture">Capture</string>
 </resources>
