+++ take-screenshot/src/main/res/values/styles.xml
